Run notebooks from dotenv on a single A100

+++ training/inference.py
@@ -33,7 +33,7 @@ def load_local_policy(
         BitsAndBytesConfig(
             load_in_4bit=True,
             bnb_4bit_quant_type="nf4",
-            bnb_4bit_compute_dtype=torch.float16,
+            bnb_4bit_compute_dtype=dtype,
             bnb_4bit_use_double_quant=True,
         )
         if load_in_4bit
